requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiohttp
requests
passlib[bcrypt]
PyJWT[crypto]
python-dotenv
